% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{assert_or_error}
\alias{assert_or_error}
\title{If condition isn't true, error}
\usage{
assert_or_error(condition, msg = "Default Message")
}
\arguments{
\item{condition}{Condition to check}

\item{msg}{Message}
}
\description{
If condition isn't true, error
}
