% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{assert_or_log}
\alias{assert_or_log}
\title{If condition isn't true, print a message at the specified log level}
\usage{
assert_or_log(condition, msg = "Default Message", level = "ERROR")
}
\arguments{
\item{condition}{Condition to check}

\item{msg}{Message}

\item{level}{}
}
\description{
If condition isn't true, print a message at the specified log level
}
