% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_text}
\alias{clean_text}
\title{Clean text}
\usage{
clean_text(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
This function cleans text by converting to lowercase, removing punctuation, and removing spaces.
}
