% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ipl_error}
\alias{ipl_error}
\title{Prints a error log message.
This causes the program to exit.}
\usage{
ipl_error(..., envir = parent.frame())
}
\description{
Prints a error log message.
This causes the program to exit.
}
