% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{ipl_log}
\alias{ipl_log}
\title{Prints a log message.}
\usage{
ipl_log(..., level, envir = parent.frame())
}
\arguments{
\item{level}{DEBUG, INFO, WARNING, ERROR}

\item{msg}{Message}
}
\description{
Will quit on ERROR.
}
